Move EngineerSpecialty to it's own .cs file
